# server/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/astrologers/featured": {
      "get": {
        "operationId": "AstrologersController_getFeaturedAstrologers",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Astrologers"
        ]
      }
    },
    "/api/v1/astrologers/top-rated": {
      "get": {
        "operationId": "AstrologersController_getTopRatedAstrologers",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Astrologers"
        ]
      }
    },
    "/api/v1/astrologers/online": {
      "get": {
        "operationId": "AstrologersController_getOnlineAstrologers",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Astrologers"
        ]
      }
    },
...
    "/api/v1/astrologers/specialization/{specialization}": {
      "get": {
        "operationId": "AstrologersController_getBySpecialization",
        "parameters": [
          {
            "name": "specialization",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Astrologers"
        ]
      }
    },
    "/api/v1/astrologers/random": {
      "get": {
        "operationId": "AstrologersController_getRandomAstrologers",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Astrologers"
        ]
      }
    },
    "/api/v1/astrologers/{astrologerId}/reviews": {
      "post": {
        "operationId": "AstrologersController_addReview",
        "parameters": [
          {
            "name": "astrologerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddReviewDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Astrologers"
        ]
      },
      "get": {
        "operationId": "AstrologersController_getAstrologerReviews",
        "parameters": [
          {
            "name": "astrologerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Astrologers"
        ]
      }
    },
    "/api/v1/astrologers/{astrologerId}/reviews/stats": {
      "get": {
        "operationId": "AstrologersController_getReviewStats",
        "parameters": [
          {
            "name": "astrologerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Astrologers"
        ]
      }
    },
    "/api/v1/astrologers/{astrologerId}": {
      "get": {
        "operationId": "AstrologersController_getAstrologerDetails",
        "parameters": [
          {
            "name": "astrologerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "includeReviews",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "reviewLimit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Astrologers"
        ]
      }
    },
...
    "/api/v1/orders/astrologer/{orderId}": {
      "get": {
        "operationId": "OrdersController_getAstrologerOrderDetails",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
...
    "/api/v1/admin/astrologers/{astrologerId}": {
      "get": {
        "operationId": "AdminAstrologersController_getAstrologerDetails",
        "parameters": [
          {
            "name": "astrologerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminAstrologers"
        ]
      },
      "delete": {
        "operationId": "AdminAstrologersController_deleteAstrologer",
        "parameters": [
          {
            "name": "astrologerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminAstrologers"
        ]
      }
    },
...
  },
  "info": {
    "title": "Project API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "AddReviewDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}